requirements.txt
black==23.1.0
requests==2.28.2
protobuf==4.21.12 --no-binary protobuf
yarl==1.8.2
pyyaml==6.0
typing_extensions==4.5.0
